remove xpub-server and move logging start up to component-app
